Extract account-loading logic into hook
- Move the account selection, cache validation, and accounts-to-load subscription out of App.tsx into a new useAccountsToLoad hook. App.tsx is simplified to import and use the hook (providing workerRef) and no longer maintains accountsToLoad or the large store subscription. The new hook encapsulates worker termination, cache checks, startTransition updates, and exposes accountsToLoad and resetAccountLoading for the rest of the app to consume.
